ffff

+++ JS.Core/Expressions/ConvertToUnsignedInteger.cs
@@ -28,7 +28,7 @@ namespace JS.Core.Expressions
             }
             else
             {
-                _tempContainer._dValue = (double)t;
+                _tempContainer._dValue = t;
                 _tempContainer._valueType = JSValueType.Double;
             }
             return _tempContainer;
